requirements.txt
# Dependencias principales
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.0.3
langchain-openai>=0.0.8

# APIs y bases de datos
openai>=1.12.0
pinecone-client>=3.0.0
python-telegram-bot>=20.7

# Utilidades
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
